refactor: delete common directory in global componenet
