+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 import LoginModal from "@/components/auth/LoginModal";
 import AppLayout from "@/layouts/AppLayout";
-import Toast from "./components/common/Toast";
+import Toast from "./components/Toast";
 import useLoginModalOpenStore from "./stores/useLoginModalOpenStore";
 
 //
