+++ src/components/header/HeaderInfoSectionSearchInput.tsx
