+++ src/layouts/AppLayout.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
-import Header from "@/components/common/header/Header";
-import Sidebar from "@/components/common/sidebar/SideBar";
+import Header from "@/components/header/Header";
+import Sidebar from "@/components/sidebar/SideBar";
 
 //
 //
